+++ src/components/mod.rs
@@ -2,8 +2,10 @@ mod b_dropdown;
 mod b_menu;
 mod b_modal;
 mod b_navbar;
+mod b_pagination;
 
 pub use b_dropdown::*;
 pub use b_menu::*;
 pub use b_modal::*;
 pub use b_navbar::*;
+pub use b_pagination::*;