+++ website/src/examples/basic_pagination.rs
@@ -1,0 +1,7 @@
+use leptos::*;
+use leptos_bulma::components::BPagination;
+
+#[component]
+pub fn BasicPagination() -> impl IntoView {
+    view! { <BPagination class="is-centered" count=10 list_size=4/> }
+}